requirements.txt
flask==3.0.3
newspaper3k
beautifulsoup4
requests
lxml
lxml_html_clean
gunicorn